restructured folders for front end
