+++ src/constants/Constants.tsx
@@ -1,5 +1,3 @@
-import { createContext } from "react";
-
 // URL
 export const LeaderBoardURL: string = "http://localhost:4000/leaderboard";
 export const BestRecordURL: string =
@@ -11,15 +9,3 @@ export const GameInstructions: string =
     "Roll until all dice are the same. Click each die to freeze it at its current value between rolls.";
 export const NoDBConnection: string = "No connection to leaderboard database.";
 
-// useContext to pass data thru component tree
-export const GameContext = createContext({
-    rollCount: 0,
-    totalTimeUsed: 0,
-    resetGame: () => {},
-});
-
-// useContext to check database connection
-export const DBConnectContext = createContext({
-    dbStatus: true,
-    setDbStatus: (dbStatus: boolean) => {},
-});
