+++ src/pages/game/Dice.tsx
